requirements.txt
streamlit==1.32.0
openai==1.12.0
transformers==4.37.0
python-dotenv==1.0.0
pydub==0.25.1
soundfile==0.12.1
scipy>=1.10.0